Finish refactoring catalog routes.
